update gradient slider
